feat: db fixture
